fixed lint errors and separate each module into a file
