+++ .eslintrc.json
@@ -1,5 +1,4 @@
 {
-    "import/no-unresolved": "off",
     "env": {
         "browser": true,
         "es6": true
